CEDS-1249 Move beforeEeach to begining of tests
Move test for empty number with type

Use translation names

Move mocks to spearate package and files

Remove override methods
