+++ test/unit/base/ControllerSpec.scala
@@ -17,56 +17,16 @@
 package unit.base
 
 import base.{MockAuthAction, MockConnectors, MockCustomsCacheService, MockExportsCacheService}
-import controllers.actions.JourneyAction
 import controllers.util.{Add, SaveAndContinue}
-import handlers.ErrorHandler
-import metrics.ExportsMetrics
-import org.mockito.ArgumentMatchers.{any, anyString}
-import org.mockito.Mockito
-import org.mockito.Mockito.when
-import org.scalatest.mockito.MockitoSugar
-import org.scalatest.{BeforeAndAfterEach, Suite}
 import play.api.libs.json.JsValue
-import play.api.mvc.Results.BadRequest
 import play.api.mvc.{AnyContentAsEmpty, AnyContentAsFormUrlEncoded, AnyContentAsJson, Request}
 import play.api.test.FakeRequest
-import play.twirl.api.Html
-import services.cache.{ExportsCacheItemBuilder, ExportsCacheModelBuilder}
+import services.cache.ExportsCacheModelBuilder
+import unit.mock.JourneyActionMocks
 import unit.tools.Stubs
 import utils.FakeRequestCSRFSupport._
 
-import scala.concurrent.{ExecutionContext, Future}
-
-trait ErrorHandlerMocks extends BeforeAndAfterEach { self: MockitoSugar with Suite =>
-
-  val mockErrorHandler: ErrorHandler = mock[ErrorHandler]
-
-  def setupErrorHandler(): Unit = {
-    when(mockErrorHandler.standardErrorTemplate(anyString, anyString, anyString)(any()))
-      .thenReturn(Html.apply(""))
-
-    when(mockErrorHandler.displayErrorPage()(any())).thenReturn(Future.successful(BadRequest(Html.apply(""))))
-  }
-
-  override protected def afterEach(): Unit = {
-    Mockito.reset(mockErrorHandler)
-    super.afterEach()
-  }
-}
-
-trait ExportsMetricsMocks extends BeforeAndAfterEach { self: MockitoSugar with Suite =>
-  val mockExportsMetrics: ExportsMetrics = mock[ExportsMetrics]
-
-  override protected def afterEach(): Unit = {
-    Mockito.reset(mockExportsMetrics)
-    super.afterEach()
-  }
-}
-
-trait JourneyActionMocks extends MockExportsCacheService with BeforeAndAfterEach { self: MockitoSugar with Suite with Stubs =>
-
-  val mockJourneyAction: JourneyAction = JourneyAction(mockExportsCacheService)(ExecutionContext.global)
-}
+import scala.concurrent.ExecutionContext
 
 trait ControllerSpec
   extends UnitSpec with Stubs with MockAuthAction with MockConnectors with MockCustomsCacheService
