+++ test/unit/controllers/CancelDeclarationControllerSpec.scala
@@ -19,14 +19,14 @@ package unit.controllers
 import com.codahale.metrics.Timer
 import controllers.CancelDeclarationController
 import forms.CancelDeclaration
-import org.mockito.ArgumentCaptor
-import org.mockito.ArgumentMatchers.{any, anyString}
-import org.mockito.Mockito.{never, verify, when}
 import forms.cancellation.CancellationChangeReason.NoLongerRequired
 import models.requests.{CancellationRequestExists, CancellationRequested, MissingDeclaration}
+import org.mockito.ArgumentMatchers.any
+import org.mockito.Mockito.when
 import play.api.libs.json.{JsValue, Json}
 import play.api.test.Helpers._
-import unit.base.{ControllerSpec, ErrorHandlerMocks, ExportsMetricsMocks}
+import unit.base.ControllerSpec
+import unit.mock.{ErrorHandlerMocks, ExportsMetricsMocks}
 import views.html.{cancel_declaration, cancellation_confirmation_page}
 
 class CancelDeclarationControllerSpec extends ControllerSpec with ErrorHandlerMocks with ExportsMetricsMocks {
