+++ test/unit/controllers/declaration/BorderTransportControllerSpec.scala
@@ -22,7 +22,8 @@ import forms.declaration.BorderTransport
 import forms.declaration.TransportCodes.{Maritime, WagonNumber}
 import play.api.libs.json.Json
 import play.api.test.Helpers._
-import unit.base.{ControllerSpec, ErrorHandlerMocks}
+import unit.base.ControllerSpec
+import unit.mock.ErrorHandlerMocks
 import views.html.declaration.border_transport
 
 class BorderTransportControllerSpec extends ControllerSpec with ErrorHandlerMocks {
